+++ packages/x6-angular-shape/src/index.ts
@@ -1,0 +1,5 @@
+import './hook'
+
+export * from './node'
+export * from './view'
+export * from './registry'