+++ sites/x6-svg-to-shape/CHANGELOG.md
@@ -1,3 +1,5 @@
+## @antv/x6-svg-to-shape [1.0.2](https://github.com/antvis/x6/compare/@antv/x6-svg-to-shape@1.0.1...@antv/x6-svg-to-shape@1.0.2) (2021-07-05)
+
 ## @antv/x6-svg-to-shape [1.0.2](https://github.com/antvis/x6/compare/@antv/x6-svg-to-shape@1.0.1...@antv/x6-svg-to-shape@1.0.2) (2021-06-23)
 
 ## @antv/x6-svg-to-shape [1.0.2](https://github.com/antvis/x6/compare/@antv/x6-svg-to-shape@1.0.1...@antv/x6-svg-to-shape@1.0.2) (2021-06-23)